requirements.txt
pyramid
pyramid_jinja2
sentry_sdk
uwsgi
